Untested DBN Gibbs sampler
This commit adds 3 methods:
 - CRBM.dbn_infer_probability
   This is a method for computing the probabilities for the hidden/pooling units in the
middle of a DBN, like in section 3.6 of the Lee paper

 - CDBN.dbn_gibbs
   This is the Gibbs sampling op.

 - CDBN._gibbs_step
   This is the body for the `tf.while_loop` in CDBN.dbn_gibbs

I feel like if there is a serious bug, it's in `dbn_infer_probability`, since I am really
confused about the way they calculate these things, and where the downsampling is happening,
etc. We need the downsampling I think but I don't want to break the network??

If there's a bug in the DBN code, it's probably something about indexing into tensorarrays
and stuff, but not a math problem.

ALSO: right now this just works on means! It does not sample yet. Their sampler doesn't
deal with the blocks/maxpooling stuff, so I didn't want to touch it yet. But the means
should be enough I think???